requirements.txt
hydra-core==1.2.0
matplotlib==3.5.2
pytorch-lightning==1.6.5
requests==2.28.1
scikit-learn==1.0.2
torch==1.12.0
transformers==4.20.1
wandb==0.12.21